Refactor(VRUI): Hack UI element size handling by reading embedded width-to-height ratio in nif name
Couldn't find a way to read node width/height data after the nif is loaded.
Didn't want to add nif parsing logic to read the vertex data in the nif.

Decided to embed the width-to-height ratio in the name of the node in the nif and then read it on the nif load and use to set the size of the element.

The way the UI nifs work is that the height is always 2 so by knowing the ratio we can set the width.
The height should never change from 2 because there is no reason to change it, just adjust the width in the nif.